requirements.txt
fastapi>=0.100.0
uvicorn>=0.24.0
python-multipart>=0.0.6
boto3>=1.26.0
pillow>=10.0.0
pydantic>=2.0.0
paddleocr>=2.7.0
pytesseract>=0.3.10
